refactor: remove legacy TRT-LLM loader code, keep only validated live transfer
Remove prototype code that was superseded by the validated live transfer path:
- Delete trtllm_loader.py (MxTrtllmSourcePublisher, MxTrtllmTargetLoader,
  manual sharding, engine build — all replaced by publish_from_worker)
- Delete trtllm_checkpoint_loader.py (CPU copy workaround, old query API)
- Delete test_trtllm_loader.py (tests for removed code)
- Remove MxLiveSource class from trtllm_live_transfer.py (fragile
  _get_torch_model — replaced by publish_from_worker which gets model
  directly in worker.setup_engine())
- Inline _build_trtllm_identity into trtllm_live_transfer.py
- Remove shape field from TensorDescriptor proto (not used by live path)
- Remove legacy exports from __init__.py

Remaining validated code:
- publish_model_params / publish_from_worker (source side)
- MxLiveWeightLoader / MxLiveCheckpointLoader (target side)
- trtllm_patches/v1.3.0rc5/ (TRT-LLM monkey patches until upstream PR)

Made-with: Cursor
