+++ modelexpress_client/python/modelexpress/__init__.py
@@ -9,7 +9,6 @@ This package provides:
 - GPUDirect Storage (GDS) for direct file-to-GPU loading
 - vLLM worker extension for serving model weights
 - Custom model loaders for FP8 model support (DeepSeek-V3, etc.)
-- TensorRT-LLM checkpoint transfer and engine building
 
 Quick Start (vLLM):
     from modelexpress import register_modelexpress_loaders
@@ -17,29 +16,12 @@ Quick Start (vLLM):
 
     # vllm serve model --load-format mx
     # Auto-detects: RDMA -> GDS -> disk
-
-Quick Start (TRT-LLM):
-    from modelexpress import MxTrtllmSourcePublisher, MxTrtllmTargetLoader
-
-    # Source: Publish checkpoint for P2P transfer
-    publisher = MxTrtllmSourcePublisher(checkpoint_dir, model_name, mx_server)
-    publisher.initialize()
-
-    # Target: Receive checkpoint and build engine
-    loader = MxTrtllmTargetLoader(model_name, mx_server, output_dir)
-    engine_dir = loader.load()
 """
 
 import logging
 
 _logger = logging.getLogger(__name__)
 _loaders_registered = False
-
-from .trtllm_loader import (
-    MxTrtllmSourcePublisher,
-    MxTrtllmTargetLoader,
-    create_trtllm_from_mx,
-)
 
 
 def register_modelexpress_loaders():
@@ -74,8 +56,4 @@ __all__ = [
     "MxClient",
     "MxGdsLoader",
     "register_modelexpress_loaders",
-    # TRT-LLM integration
-    "MxTrtllmSourcePublisher",
-    "MxTrtllmTargetLoader",
-    "create_trtllm_from_mx",
 ]
