--- modelexpress_client/python/modelexpress/trtllm_loader.py
@@ -1,1328 +1,0 @@
-# SPDX-FileCopyrightText: Copyright (c) 2025-2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
-# SPDX-License-Identifier: Apache-2.0
-
-"""
-ModelExpress TensorRT-LLM Integration.
-
-Enables P2P weight transfer for TRT-LLM via NIXL RDMA.
-
-Supports TWO modes:
-
-1. **HuggingFace Mode** (recommended for DeepSeek-V3/R1, Llama, Qwen):
-   - Source loads HuggingFace model directly to GPU
-   - Transfers raw weights via RDMA (similar to vLLM P2P)
-   - Target uses TRT-LLM's PyTorch backend for direct inference
-   - No checkpoint conversion or engine build required!
-
-2. **Checkpoint Mode** (traditional TRT-LLM workflow):
-   - Requires pre-converted TRT-LLM checkpoint (config.json + rank*.safetensors)
-   - Transfers checkpoint weights via RDMA
-   - Target optionally builds TensorRT engine
-
-This module provides:
-    - MxTrtllmSourcePublisher: Loads HuggingFace model or checkpoint, registers with NIXL
-    - MxTrtllmTargetLoader: Receives weights via RDMA, uses with TRT-LLM
-
-Usage (HuggingFace mode - recommended):
-    # Source side - load HuggingFace model and publish for P2P
-    publisher = MxTrtllmSourcePublisher(
-        hf_model_path="/path/to/deepseek-v3",  # HuggingFace format
-        model_name="deepseek-v3",
-        mx_server="modelexpress-server:8001",
-        tp_size=8,
-    )
-    publisher.initialize()
-    
-    # Target side - receive weights and use directly with TRT-LLM
-    loader = MxTrtllmTargetLoader(
-        model_name="deepseek-v3",
-        mx_server="modelexpress-server:8001",
-        output_dir="/tmp/mx_trtllm",
-    )
-    model_path = loader.load(use_pytorch_backend=True)
-    
-    # Use with TRT-LLM PyTorch backend
-    from tensorrt_llm import LLM
-    llm = LLM(model=model_path, backend="pytorch")
-
-Usage (Checkpoint mode - traditional):
-    # Source with pre-converted checkpoint
-    publisher = MxTrtllmSourcePublisher(
-        checkpoint_dir="/path/to/trtllm-checkpoint",
-        model_name="llama-70b",
-        mx_server="modelexpress-server:8001"
-    )
-    
-    # Target receives and builds engine
-    loader = MxTrtllmTargetLoader(...)
-    engine_dir = loader.load(skip_build=False)
-"""
-
-from __future__ import annotations
-
-import json
-import logging
-import os
-import subprocess
-import time
-import uuid
-from pathlib import Path
-from typing import TYPE_CHECKING
-
-import torch
-
-from .client import MxClient
-from . import p2p_pb2
-
-if TYPE_CHECKING:
-    from .nixl_transfer import NixlTransferManager
-
-logger = logging.getLogger("modelexpress.trtllm_loader")
-
-
-def _parse_server_address(address: str) -> str:
-    """Strip http:// or https:// prefix from server address."""
-    if address.startswith("http://"):
-        return address[7:]
-    elif address.startswith("https://"):
-        return address[8:]
-    return address
-
-
-def _build_trtllm_identity(
-    model_name: str,
-    tp_size: int = 1,
-    ep_size: int = 1,
-    dtype: str = "bfloat16",
-) -> p2p_pb2.SourceIdentity:
-    """Build a SourceIdentity for TRT-LLM weight transfers."""
-    from importlib.metadata import version as pkg_version
-
-    try:
-        mx_version = pkg_version("modelexpress")
-    except Exception:
-        mx_version = "0.0.0"
-
-    return p2p_pb2.SourceIdentity(
-        mx_version=mx_version,
-        mx_source_type=p2p_pb2.MX_SOURCE_TYPE_WEIGHTS,
-        model_name=model_name,
-        backend_framework=p2p_pb2.BACKEND_FRAMEWORK_TRT_LLM,
-        tensor_parallel_size=tp_size,
-        pipeline_parallel_size=1,
-        expert_parallel_size=ep_size,
-        dtype=dtype,
-    )
-
-
-class MxTrtllmSourcePublisher:
-    """
-    Publishes TRT-LLM weights for P2P transfer.
-    
-    This class supports three modes:
-    
-    1. **HuggingFace mode** (recommended for DeepSeek, Llama, Qwen):
-       Loads HuggingFace model directly to GPU using TRT-LLM's PyTorch backend.
-       Transfers raw weights via RDMA - fastest path, no conversion needed.
-       
-    2. **Checkpoint mode**: Loads TRT-LLM checkpoint files (safetensors) to GPU.
-       Target receives checkpoint weights.
-       
-    3. **Engine mode**: Loads pre-built TensorRT engine's weights to GPU.
-       Requires source and target to have same GPU architecture.
-    
-    The weights remain in GPU memory for the lifetime of this object,
-    allowing multiple targets to receive them.
-    
-    Args:
-        checkpoint_dir: Path to TRT-LLM checkpoint directory (for checkpoint mode)
-        model_name: Model identifier for coordination
-        mx_server: ModelExpress server address (host:port)
-        engine_dir: Optional path to pre-built engine directory
-        hf_model_path: Path to HuggingFace model (for HuggingFace mode)
-        tp_size: Tensor parallelism size (for HuggingFace mode)
-        dtype: Model dtype (default: "bfloat16")
-    """
-    
-    def __init__(
-        self,
-        model_name: str,
-        mx_server: str = "modelexpress-server:8001",
-        checkpoint_dir: str | None = None,
-        engine_dir: str | None = None,
-        hf_model_path: str | None = None,
-        tp_size: int = 1,
-        ep_size: int = 1,
-        num_experts: int = 0,
-        dtype: str = "bfloat16",
-    ):
-        self.model_name = model_name
-        self.mx_server = _parse_server_address(mx_server)
-        self.checkpoint_dir = Path(checkpoint_dir) if checkpoint_dir else None
-        self.engine_dir = Path(engine_dir) if engine_dir else None
-        self.hf_model_path = Path(hf_model_path) if hf_model_path else None
-        self.tp_size = tp_size
-        self.ep_size = ep_size
-        self.num_experts = num_experts
-        self.dtype = dtype
-        self.config: dict = {}
-        self.nixl_managers: dict[int, NixlTransferManager] = {}
-        self.weights: dict[int, dict[str, torch.Tensor]] = {}
-        self._initialized = False
-        self._engine_bytes: bytes | None = None
-        self._mode: str = "unknown"
-        self._mx_client = MxClient(server_url=mx_server)
-        self._worker_id = uuid.uuid4().hex[:8]
-        self._identity = _build_trtllm_identity(
-            model_name=model_name, tp_size=tp_size, ep_size=ep_size, dtype=dtype,
-        )
-        
-    def initialize(self, build_if_needed: bool = True) -> None:
-        """
-        Load model/checkpoint/engine and register with NIXL.
-        
-        This method auto-detects the mode based on provided paths:
-        1. If hf_model_path provided: HuggingFace mode (recommended)
-        2. If engine_dir provided: Engine mode
-        3. If checkpoint_dir provided: Checkpoint mode
-        
-        Args:
-            build_if_needed: If True and checkpoint mode with no engine, build engine
-        """
-        from .nixl_transfer import NixlTransferManager, is_nixl_available
-        
-        if not is_nixl_available():
-            raise RuntimeError("NIXL is not available. Install with: pip install nixl[cu12]")
-        
-        if self._initialized:
-            logger.warning("Already initialized, skipping")
-            return
-        
-        logger.info(f"Initializing TRT-LLM source publisher for {self.model_name}")
-        
-        # Auto-detect mode
-        if self.hf_model_path and self.hf_model_path.exists():
-            self._mode = "huggingface"
-            self._initialize_huggingface_mode()
-        elif self.checkpoint_dir and self.checkpoint_dir.exists():
-            self._mode = "checkpoint"
-            self._initialize_checkpoint_mode(build_if_needed)
-        else:
-            raise ValueError(
-                "Must provide either hf_model_path (HuggingFace model) or "
-                "checkpoint_dir (TRT-LLM checkpoint)"
-            )
-        
-        # Publish to MX server
-        self._publish_to_mx_server()
-        
-        self._initialized = True
-        logger.info(f"TRT-LLM source publisher initialized ({self._mode} mode)")
-    
-    def _initialize_huggingface_mode(self) -> None:
-        """Load HuggingFace model directly to GPU (like vLLM P2P)."""
-        from .nixl_transfer import NixlTransferManager
-        
-        logger.info(f"HuggingFace mode: Loading {self.hf_model_path} with TP={self.tp_size}")
-        
-        # Store HF config for metadata
-        hf_config_path = self.hf_model_path / "config.json"
-        if hf_config_path.exists():
-            with open(hf_config_path) as f:
-                hf_config = json.load(f)
-
-            if self.num_experts == 0:
-                self.num_experts = hf_config.get("n_routed_experts", hf_config.get("num_local_experts", 0))
-
-            self.config = {
-                "architecture": hf_config.get("model_type", "unknown"),
-                "hidden_size": hf_config.get("hidden_size"),
-                "num_layers": hf_config.get("num_hidden_layers"),
-                "num_experts": self.num_experts,
-                "ep_size": self.ep_size,
-                "dtype": self.dtype,
-                "mapping": {
-                    "world_size": self.tp_size * self.ep_size,
-                    "tp_size": self.tp_size,
-                    "ep_size": self.ep_size,
-                    "pp_size": 1,
-                },
-                "_hf_model_path": str(self.hf_model_path),
-                "_mode": "huggingface",
-            }
-            logger.info(
-                "Model: %s, hidden_size=%s, num_experts=%d, ep_size=%d",
-                self.config["architecture"], self.config["hidden_size"],
-                self.num_experts, self.ep_size,
-            )
-        
-        # Load model weights for each rank.
-        # With EP, total ranks = tp_size * ep_size. Each EP rank gets a subset of experts.
-        # With pure EP (tp_size=1), rank == ep_rank and device_id == ep_rank.
-        total_ranks = self.tp_size * self.ep_size if self.ep_size > 1 else self.tp_size
-        total_bytes = 0
-
-        for rank in range(total_ranks):
-            ep_rank = rank // self.tp_size if self.ep_size > 1 else 0
-            tp_rank = rank % self.tp_size if self.tp_size > 1 else 0
-            device_id = rank % torch.cuda.device_count()
-
-            logger.info(
-                "Loading rank %d/%d (tp=%d, ep=%d) to GPU %d...",
-                rank, total_ranks, tp_rank, ep_rank, device_id,
-            )
-
-            weights = self._load_hf_weights_for_rank(
-                rank, tp_rank=tp_rank, ep_rank=ep_rank
-            )
-            self.weights[rank] = weights
-
-            rank_bytes = sum(t.numel() * t.element_size() for t in weights.values())
-            total_bytes += rank_bytes
-            logger.info("Rank %d: %d tensors, %.2f GB", rank, len(weights), rank_bytes / 1e9)
-
-            agent_name = f"trtllm-hf-source-rank{rank}-{uuid.uuid4().hex[:8]}"
-            nixl_manager = NixlTransferManager(
-                agent_name=agent_name,
-                device_id=device_id,
-            )
-            nixl_manager.initialize()
-            nixl_manager.register_tensors(weights)
-            self.nixl_managers[rank] = nixl_manager
-            logger.info("Rank %d: Registered with NIXL", rank)
-
-        logger.info("HuggingFace mode: %.2f GB across %d ranks", total_bytes / 1e9, total_ranks)
-    
-    def _load_hf_weights_for_rank(
-        self, rank: int, tp_rank: int = 0, ep_rank: int = 0
-    ) -> dict[str, torch.Tensor]:
-        """
-        Load HuggingFace model weights for a specific rank.
-
-        Handles both TP sharding and MoE expert-parallel partitioning.
-        For EP: filters expert weights to only include this rank's assigned experts.
-        For TP: shards attention/MLP weights along appropriate dimensions.
-        """
-        try:
-            import safetensors.torch
-        except ImportError:
-            raise ImportError("safetensors required. Install with: pip install safetensors") from None
-
-        weights: dict[str, torch.Tensor] = {}
-
-        safetensor_files = list(self.hf_model_path.glob("*.safetensors"))
-        if not safetensor_files:
-            raise FileNotFoundError(f"No safetensors files found in {self.hf_model_path}")
-
-        logger.info("Found %d safetensors files", len(safetensor_files))
-
-        device_id = rank % torch.cuda.device_count()
-        device = f"cuda:{device_id}"
-
-        dtype_map = {
-            "bfloat16": torch.bfloat16,
-            "float16": torch.float16,
-            "float32": torch.float32,
-        }
-        target_dtype = dtype_map.get(self.dtype, torch.bfloat16)
-
-        # Compute which experts belong to this EP rank
-        assigned_experts: set[int] | None = None
-        if self.ep_size > 1 and self.num_experts > 0:
-            experts_per_rank = self.num_experts // self.ep_size
-            start = ep_rank * experts_per_rank
-            assigned_experts = set(range(start, start + experts_per_rank))
-            logger.info(
-                "EP rank %d: assigned experts %d-%d (%d of %d)",
-                ep_rank, start, start + experts_per_rank - 1,
-                experts_per_rank, self.num_experts,
-            )
-
-        import re
-
-        for sf_path in safetensor_files:
-            file_weights = safetensors.torch.load_file(str(sf_path), device=device)
-
-            for name, tensor in file_weights.items():
-                if tensor.dtype != target_dtype and tensor.dtype in [torch.float32, torch.float16, torch.bfloat16]:
-                    tensor = tensor.to(target_dtype)
-
-                # MoE expert filtering: skip experts not assigned to this EP rank
-                if assigned_experts is not None:
-                    expert_match = re.search(r"\.experts\.(\d+)\.", name)
-                    if expert_match:
-                        expert_idx = int(expert_match.group(1))
-                        if expert_idx not in assigned_experts:
-                            continue
-
-                # TP sharding for non-expert weights (attention, etc.)
-                if self.tp_size > 1:
-                    # Don't TP-shard individual expert weights (they're EP-partitioned)
-                    if ".experts." not in name:
-                        tensor = self._shard_tensor_for_rank(name, tensor, tp_rank)
-
-                if tensor is not None:
-                    weights[name] = tensor
-
-        return weights
-    
-    def _shard_tensor_for_rank(
-        self, name: str, tensor: torch.Tensor, rank: int
-    ) -> torch.Tensor | None:
-        """
-        Shard a tensor for tensor parallelism.
-        
-        This implements basic TP sharding based on tensor name patterns.
-        For more complex models, override this method.
-        """
-        if self.tp_size == 1:
-            return tensor
-        
-        # Common patterns for column-parallel (shard dim=-1) and row-parallel (shard dim=0)
-        col_parallel_patterns = [
-            "q_proj", "k_proj", "v_proj", "gate_proj", "up_proj",
-            "qkv_proj", "gate_up_proj", "query_key_value",
-        ]
-        row_parallel_patterns = [
-            "o_proj", "down_proj", "out_proj", "dense",
-        ]
-        
-        # Check if tensor should be sharded
-        shard_dim = None
-        for pattern in col_parallel_patterns:
-            if pattern in name.lower():
-                shard_dim = -1  # Shard last dimension
-                break
-        
-        if shard_dim is None:
-            for pattern in row_parallel_patterns:
-                if pattern in name.lower():
-                    shard_dim = 0  # Shard first dimension
-                    break
-        
-        if shard_dim is None:
-            # Not a tensor that needs sharding, return as-is
-            return tensor
-        
-        # Perform sharding
-        dim_size = tensor.shape[shard_dim]
-        if dim_size % self.tp_size != 0:
-            # Can't evenly shard, return full tensor
-            return tensor
-        
-        shard_size = dim_size // self.tp_size
-        start_idx = rank * shard_size
-        end_idx = start_idx + shard_size
-        
-        if shard_dim == -1:
-            return tensor[..., start_idx:end_idx].contiguous()
-        else:
-            return tensor[start_idx:end_idx, ...].contiguous()
-    
-    def _initialize_checkpoint_mode(self, build_if_needed: bool) -> None:
-        """Load TRT-LLM checkpoint (traditional mode)."""
-        logger.info(f"Checkpoint mode: Loading from {self.checkpoint_dir}")
-        
-        # Parse config.json
-        config_path = self.checkpoint_dir / "config.json"
-        if not config_path.exists():
-            raise FileNotFoundError(f"Config not found: {config_path}")
-        
-        with open(config_path) as f:
-            self.config = json.load(f)
-        
-        self.config["_mode"] = "checkpoint"
-        
-        world_size = self.config.get("mapping", {}).get("world_size", 1)
-        tp_size = self.config.get("mapping", {}).get("tp_size", 1)
-        pp_size = self.config.get("mapping", {}).get("pp_size", 1)
-        
-        logger.info(f"Config: world_size={world_size}, tp_size={tp_size}, pp_size={pp_size}")
-        
-        # Check for engine or build if needed
-        if self.engine_dir and self.engine_dir.exists():
-            logger.info(f"Using pre-built engine from {self.engine_dir}")
-            self._load_engine_for_transfer()
-        elif build_if_needed:
-            logger.info("No engine found, building from checkpoint...")
-            self._build_engine()
-            self._load_engine_for_transfer()
-        else:
-            logger.info("Checkpoint-only mode (no engine transfer)")
-        
-        # Load and register each rank's weights
-        total_bytes = 0
-        for rank in range(world_size):
-            weights_path = self.checkpoint_dir / f"rank{rank}.safetensors"
-            if not weights_path.exists():
-                raise FileNotFoundError(f"Weights not found: {weights_path}")
-            
-            logger.info(f"Loading rank {rank} weights from {weights_path}")
-            weights = self._load_safetensors_to_gpu(weights_path, rank)
-            self.weights[rank] = weights
-            
-            rank_bytes = sum(t.numel() * t.element_size() for t in weights.values())
-            total_bytes += rank_bytes
-            logger.info(f"Rank {rank}: {len(weights)} tensors, {rank_bytes / 1e9:.2f} GB")
-            
-            agent_name = f"trtllm-source-rank{rank}-{uuid.uuid4().hex[:8]}"
-            nixl_manager = NixlTransferManager(
-                agent_name=agent_name,
-                device_id=rank
-            )
-            nixl_manager.initialize()
-            nixl_manager.register_tensors(weights)
-            self.nixl_managers[rank] = nixl_manager
-            logger.info(f"Rank {rank}: Registered {len(weights)} tensors with NIXL")
-        
-        logger.info(f"Checkpoint mode: {total_bytes / 1e9:.2f} GB across {world_size} ranks")
-    
-    def _build_engine(self) -> None:
-        """Build TRT-LLM engine from checkpoint."""
-        import subprocess
-        
-        if self.engine_dir is None:
-            self.engine_dir = self.checkpoint_dir.parent / "engine"
-        
-        self.engine_dir.mkdir(parents=True, exist_ok=True)
-        
-        logger.info(f"Building engine from {self.checkpoint_dir} to {self.engine_dir}")
-        
-        cmd = [
-            "trtllm-build",
-            "--checkpoint_dir", str(self.checkpoint_dir),
-            "--output_dir", str(self.engine_dir),
-            "--gemm_plugin", "auto",
-            "--max_batch_size", "8",
-            "--max_input_len", "2048", 
-            "--max_seq_len", "4096",
-        ]
-        
-        logger.info(f"Running: {' '.join(cmd)}")
-        
-        t0 = time.time()
-        result = subprocess.run(cmd, capture_output=True, text=True)
-        build_time = time.time() - t0
-        
-        if result.returncode != 0:
-            logger.error(f"Engine build failed:\n{result.stderr}")
-            raise RuntimeError(f"trtllm-build failed: {result.stderr}")
-        
-        logger.info(f"Engine built in {build_time:.1f}s")
-    
-    def _load_engine_for_transfer(self) -> None:
-        """Load engine file for transfer to targets."""
-        if self.engine_dir is None:
-            return
-        
-        # Find engine file (usually rank0.engine or similar)
-        engine_files = list(self.engine_dir.glob("*.engine"))
-        if not engine_files:
-            logger.warning("No .engine files found in engine directory")
-            return
-        
-        # Read engine bytes for transfer
-        engine_file = engine_files[0]
-        logger.info(f"Loading engine file: {engine_file}")
-        
-        with open(engine_file, "rb") as f:
-            self._engine_bytes = f.read()
-        
-        logger.info(f"Engine loaded: {len(self._engine_bytes) / 1e6:.1f} MB")
-    
-    def _load_safetensors_to_gpu(
-        self, path: Path, device_id: int
-    ) -> dict[str, torch.Tensor]:
-        """Load safetensors file directly to GPU memory."""
-        try:
-            import safetensors.torch
-        except ImportError:
-            raise ImportError("safetensors not installed. Install with: pip install safetensors") from None
-        
-        with torch.cuda.device(device_id):
-            weights = safetensors.torch.load_file(
-                str(path), device=f"cuda:{device_id}"
-            )
-        return weights
-    
-    def _publish_to_mx_server(self) -> None:
-        """Publish tensor metadata to ModelExpress server (per-worker)."""
-        for rank, nixl_manager in self.nixl_managers.items():
-            weights = self.weights[rank]
-
-            tensor_protos = [
-                p2p_pb2.TensorDescriptor(
-                    name=name,
-                    addr=tensor.data_ptr(),
-                    size=tensor.numel() * tensor.element_size(),
-                    device_id=rank,
-                    dtype=str(tensor.dtype),
-                    shape=list(tensor.shape),
-                )
-                for name, tensor in weights.items()
-            ]
-
-            worker = p2p_pb2.WorkerMetadata(
-                worker_rank=rank,
-                nixl_metadata=nixl_manager.nixl_metadata,
-                tensors=tensor_protos,
-            )
-
-            mx_source_id = self._mx_client.publish_metadata(
-                identity=self._identity,
-                worker=worker,
-                worker_id=self._worker_id,
-            )
-            logger.info(
-                f"Rank {rank}: Published metadata (mx_source_id={mx_source_id}, "
-                f"worker_id={self._worker_id}, {len(tensor_protos)} tensors)"
-            )
-
-        logger.info(
-            f"Published metadata for {len(self.nixl_managers)} workers "
-            f"to ModelExpress server"
-        )
-    
-    def get_engine_bytes(self) -> bytes | None:
-        """Get the serialized engine bytes for transfer."""
-        return self._engine_bytes
-    
-    def shutdown(self) -> None:
-        """Clean up NIXL resources and GPU memory."""
-        logger.info("Shutting down TRT-LLM source publisher")
-
-        for rank, nixl_manager in self.nixl_managers.items():
-            nixl_manager.shutdown()
-            logger.debug(f"Rank {rank}: NIXL manager shutdown")
-
-        self.nixl_managers.clear()
-        self.weights.clear()
-        self._initialized = False
-        self._mx_client.close()
-
-        import gc
-        gc.collect()
-        torch.cuda.empty_cache()
-
-        logger.info("TRT-LLM source publisher shutdown complete")
-    
-    def __enter__(self):
-        self.initialize()
-        return self
-    
-    def __exit__(self, exc_type, exc_val, exc_tb):
-        self.shutdown()
-        return False
-
-
-class MxTrtllmTargetLoader:
-    """
-    Receives TRT-LLM weights via P2P and prepares for inference.
-    
-    Supports two modes based on source:
-    
-    1. **HuggingFace mode** (when source used hf_model_path):
-       - Receives raw HuggingFace weights via RDMA
-       - Saves in HuggingFace format for TRT-LLM PyTorch backend
-       - No engine build required - use directly with LLM(backend="pytorch")
-       
-    2. **Checkpoint mode** (when source used checkpoint_dir):
-       - Receives TRT-LLM checkpoint weights via RDMA
-       - Optionally builds TensorRT engine
-    
-    Args:
-        model_name: Model identifier in ModelExpress
-        mx_server: ModelExpress server address (host:port)
-        output_dir: Local directory for received weights
-        build_config: Optional trtllm-build configuration (checkpoint mode only)
-        tp_size: Expected tensor parallelism size
-    """
-    
-    def __init__(
-        self,
-        model_name: str,
-        mx_server: str = "modelexpress-server:8001",
-        output_dir: str = "/tmp/mx_trtllm",
-        build_config: dict | None = None,
-        tp_size: int = 1,
-    ):
-        self.model_name = model_name
-        self.mx_server = _parse_server_address(mx_server)
-        self.output_dir = Path(output_dir)
-        self.checkpoint_dir = self.output_dir / "checkpoint"
-        self.engine_dir = self.output_dir / "engine"
-        self.hf_output_dir = self.output_dir / "hf_model"
-        self.tp_size = tp_size
-
-        self.build_config = build_config or {
-            "gemm_plugin": "auto",
-            "max_batch_size": "8",
-            "max_input_len": "2048",
-            "max_seq_len": "4096",
-        }
-
-        self.config: dict = {}
-        self.transfer_stats: dict = {}
-        self._source_mode: str = "unknown"
-        self._mx_client = MxClient(server_url=mx_server)
-        self._identity = _build_trtllm_identity(
-            model_name=model_name, tp_size=tp_size,
-        )
-    
-    def load(
-        self,
-        skip_build: bool = False,
-        use_pytorch_backend: bool = False,
-    ) -> str:
-        """
-        Receive weights via P2P and prepare for inference.
-        
-        Behavior depends on source mode:
-        - HuggingFace mode: Returns path to HF model for PyTorch backend
-        - Checkpoint mode: Returns path to checkpoint or engine
-        
-        Args:
-            skip_build: If True, skip engine build (checkpoint mode only)
-            use_pytorch_backend: If True, prepare for TRT-LLM PyTorch backend
-            
-        Returns:
-            Path to model directory suitable for TRT-LLM LLM()
-            
-        Example:
-            # For HuggingFace/PyTorch backend mode
-            model_path = loader.load(use_pytorch_backend=True)
-            from tensorrt_llm import LLM
-            llm = LLM(model=model_path, backend="pytorch")
-            
-            # For traditional engine mode
-            engine_path = loader.load(skip_build=False)
-            llm = LLM(model=engine_path)
-        """
-        logger.info(f"Loading {self.model_name} via ModelExpress P2P transfer")
-        
-        # 1. Query MX server for source
-        source_meta = self._query_source()
-        
-        # 2. Detect source mode from config
-        self._source_mode = self.config.get("_mode", "checkpoint")
-        logger.info(f"Source mode: {self._source_mode}")
-        
-        # 3. Receive weights via NIXL
-        if self._source_mode == "huggingface":
-            self.hf_output_dir.mkdir(parents=True, exist_ok=True)
-            self._receive_hf_weights(source_meta)
-            
-            # Save config for TRT-LLM
-            self._save_hf_config()
-            
-            logger.info(f"HuggingFace model saved to {self.hf_output_dir}")
-            return str(self.hf_output_dir)
-        else:
-            # Checkpoint mode
-            self.checkpoint_dir.mkdir(parents=True, exist_ok=True)
-            self._receive_checkpoint(source_meta)
-            
-            # Save config from source metadata for TRT-LLM
-            config_path = self.checkpoint_dir / "config.json"
-            if self.config and not config_path.exists():
-                # Create config.json from source metadata
-                self._save_checkpoint_config()
-            elif config_path.exists():
-                with open(config_path) as f:
-                    self.config = json.load(f)
-                logger.info("Loaded config from checkpoint")
-            
-            if skip_build or use_pytorch_backend:
-                logger.info(f"Checkpoint saved to {self.checkpoint_dir}")
-                return str(self.checkpoint_dir)
-            
-            # Build engine locally
-            engine_path = self._build_engine()
-            logger.info(f"Engine built at {engine_path}")
-            return engine_path
-    
-    def _receive_hf_weights(self, source_meta) -> None:
-        """Receive HuggingFace format weights via NIXL."""
-        from .nixl_transfer import NixlTransferManager
-        from .types import TensorDescriptor
-        
-        total_bytes = 0
-        total_time = 0
-        all_weights: dict[str, torch.Tensor] = {}
-        
-        for worker in source_meta.workers:
-            rank = worker.worker_rank
-            tensor_count = len(worker.tensors)
-            
-            logger.info(f"Receiving rank {rank}: {tensor_count} tensors")
-            
-            agent_name = f"trtllm-hf-target-rank{rank}-{uuid.uuid4().hex[:8]}"
-            nixl_manager = NixlTransferManager(
-                agent_name=agent_name,
-                device_id=rank
-            )
-            nixl_manager.initialize()
-            
-            # Allocate GPU tensors to receive into
-            weights = self._allocate_tensors(worker.tensors, rank)
-            nixl_manager.register_tensors(weights)
-            
-            # Build source descriptors
-            source_tensors = [
-                TensorDescriptor(
-                    name=t.name,
-                    addr=t.addr,
-                    size=t.size,
-                    device_id=t.device_id,
-                    dtype=t.dtype,
-                )
-                for t in worker.tensors
-            ]
-            
-            # Receive via RDMA
-            t0 = time.perf_counter()
-            bytes_received, _, _duration = nixl_manager.receive_from_source(
-                source_metadata=worker.nixl_metadata,
-                source_tensors=source_tensors,
-                timeout_seconds=600,
-                coalesce_transfers=False,
-            )
-            transfer_time = time.perf_counter() - t0
-            
-            total_bytes += bytes_received
-            total_time += transfer_time
-            
-            bandwidth = (bytes_received * 8) / (transfer_time * 1e9) if transfer_time > 0 else 0
-            logger.info(
-                f"Rank {rank}: Received {bytes_received / 1e9:.2f} GB "
-                f"in {transfer_time:.2f}s ({bandwidth:.1f} Gbps)"
-            )
-            
-            # Collect all weights
-            all_weights.update(weights)
-            
-            # Cleanup NIXL
-            nixl_manager.shutdown()
-        
-        # Save weights in HuggingFace format (single safetensors file)
-        self._save_hf_weights(all_weights)
-        
-        # Store transfer stats
-        self.transfer_stats = {
-            "total_bytes": total_bytes,
-            "total_time": total_time,
-            "bandwidth_gbps": (total_bytes * 8) / (total_time * 1e9) if total_time > 0 else 0,
-        }
-        
-        logger.info(
-            f"HF Transfer complete: {total_bytes / 1e9:.2f} GB total "
-            f"in {total_time:.2f}s ({self.transfer_stats['bandwidth_gbps']:.1f} Gbps)"
-        )
-    
-    def _save_hf_weights(self, weights: dict[str, torch.Tensor]) -> None:
-        """Save weights in HuggingFace safetensors format."""
-        try:
-            import safetensors.torch
-        except ImportError:
-            raise ImportError("safetensors required. Install with: pip install safetensors") from None
-        
-        # Move to CPU for saving
-        cpu_weights = {k: v.cpu() for k, v in weights.items()}
-        
-        weights_path = self.hf_output_dir / "model.safetensors"
-        safetensors.torch.save_file(cpu_weights, str(weights_path))
-        
-        total_size = sum(t.numel() * t.element_size() for t in cpu_weights.values())
-        logger.info(f"Saved {len(cpu_weights)} tensors ({total_size / 1e9:.2f} GB) to {weights_path}")
-        
-        # Free GPU memory
-        del weights
-        torch.cuda.empty_cache()
-    
-    def _save_hf_config(self) -> None:
-        """Save HuggingFace config for TRT-LLM."""
-        # Extract HF config from source metadata
-        hf_config = {
-            "architectures": [self.config.get("architecture", "AutoModelForCausalLM")],
-            "model_type": self.config.get("architecture", "llama"),
-            "hidden_size": self.config.get("hidden_size", 4096),
-            "num_hidden_layers": self.config.get("num_layers", 32),
-            "torch_dtype": self.config.get("dtype", "bfloat16"),
-        }
-        
-        config_path = self.hf_output_dir / "config.json"
-        with open(config_path, "w") as f:
-            json.dump(hf_config, f, indent=2)
-        
-        logger.info(f"Saved HF config to {config_path}")
-    
-    def _save_checkpoint_config(self) -> None:
-        """Save config.json for TRT-LLM checkpoint format."""
-        # For HuggingFace source mode, create a HuggingFace-compatible config
-        # that TRT-LLM's PyTorch backend can understand
-        config = {
-            "architectures": [self.config.get("architecture", "LlamaForCausalLM")],
-            "model_type": self.config.get("architecture", "llama"),
-            "hidden_size": self.config.get("hidden_size", 4096),
-            "num_hidden_layers": self.config.get("num_layers", 32),
-            "torch_dtype": self.config.get("dtype", "bfloat16"),
-            # Include mapping info for TRT-LLM
-            "tensor_parallel_size": self.config.get("mapping", {}).get("tp_size", self.tp_size),
-        }
-        
-        # If source was HuggingFace, get the original path for reference
-        if self.config.get("_hf_model_path"):
-            config["_source_model"] = self.config.get("_hf_model_path")
-        
-        config_path = self.checkpoint_dir / "config.json"
-        with open(config_path, "w") as f:
-            json.dump(config, f, indent=2)
-        
-        logger.info(f"Saved checkpoint config to {config_path}")
-    
-    def _query_source(self, timeout: int = 3600):
-        """Query ModelExpress server for source metadata.
-
-        Uses ListSources to find READY instances, then GetMetadata per
-        worker to fetch full tensor descriptors.  Returns a namespace
-        with a ``.workers`` list so callers iterate the same way as before.
-        """
-        import grpc
-
-        logger.info(f"Querying for source model: {self.model_name}")
-
-        start = time.time()
-        poll_interval = 30
-
-        while time.time() - start < timeout:
-            try:
-                list_resp = self._mx_client.list_sources(
-                    identity=self._identity,
-                    status_filter=p2p_pb2.SOURCE_STATUS_READY,
-                )
-                if list_resp.instances:
-                    workers = []
-                    for inst in list_resp.instances:
-                        meta_resp = self._mx_client.get_metadata(
-                            mx_source_id=inst.mx_source_id,
-                            worker_id=inst.worker_id,
-                        )
-                        if meta_resp.found and meta_resp.worker.tensors:
-                            workers.append(meta_resp.worker)
-
-                    if workers:
-                        logger.info(f"Found source with {len(workers)} workers")
-                        self.config = {}
-
-                        class _SourceMeta:
-                            pass
-
-                        result = _SourceMeta()
-                        result.workers = workers
-                        return result
-
-                elapsed = int(time.time() - start)
-                logger.info(f"Source not ready, waiting... ({elapsed}s / {timeout}s)")
-
-            except grpc.RpcError as e:
-                logger.warning(f"gRPC error: {e}, retrying...")
-
-            time.sleep(poll_interval)
-
-        raise TimeoutError(f"Source not found for {self.model_name} after {timeout}s")
-    
-    def _receive_checkpoint(self, source_meta) -> None:
-        """Receive checkpoint weights via NIXL and reconstruct full HF format."""
-        from .nixl_transfer import NixlTransferManager
-        from .types import TensorDescriptor
-        
-        total_bytes = 0
-        total_time = 0
-        
-        # Collect all weights from all ranks for reconstruction
-        all_rank_weights: dict[int, dict[str, torch.Tensor]] = {}
-        
-        for worker in source_meta.workers:
-            rank = worker.worker_rank
-            tensor_count = len(worker.tensors)
-            
-            logger.info(f"Receiving rank {rank}: {tensor_count} tensors")
-            
-            agent_name = f"trtllm-target-rank{rank}-{uuid.uuid4().hex[:8]}"
-            nixl_manager = NixlTransferManager(
-                agent_name=agent_name,
-                device_id=rank
-            )
-            nixl_manager.initialize()
-            
-            # Allocate GPU tensors to receive into
-            weights = self._allocate_tensors(worker.tensors, rank)
-            nixl_manager.register_tensors(weights)
-            
-            # Build source descriptors
-            source_tensors = [
-                TensorDescriptor(
-                    name=t.name,
-                    addr=t.addr,
-                    size=t.size,
-                    device_id=t.device_id,
-                    dtype=t.dtype,
-                )
-                for t in worker.tensors
-            ]
-            
-            # Receive via RDMA
-            t0 = time.perf_counter()
-            bytes_received, _, _duration = nixl_manager.receive_from_source(
-                source_metadata=worker.nixl_metadata,
-                source_tensors=source_tensors,
-                timeout_seconds=600,  # 10 minute timeout per rank
-                coalesce_transfers=False,  # Disabled for cross-node RDMA compatibility
-            )
-            transfer_time = time.perf_counter() - t0
-            
-            total_bytes += bytes_received
-            total_time += transfer_time
-            
-            bandwidth = (bytes_received * 8) / (transfer_time * 1e9) if transfer_time > 0 else 0
-            logger.info(
-                f"Rank {rank}: Received {bytes_received / 1e9:.2f} GB "
-                f"in {transfer_time:.2f}s ({bandwidth:.1f} Gbps)"
-            )
-            
-            # Copy weights to CPU immediately to avoid GPU memory issues
-            # when switching between CUDA devices
-            cpu_weights = {k: v.cpu() for k, v in weights.items()}
-            all_rank_weights[rank] = cpu_weights
-            logger.info(f"  Copied {len(cpu_weights)} tensors to CPU")
-            
-            # Free GPU memory
-            del weights
-            torch.cuda.empty_cache()
-            
-            # Cleanup NIXL
-            nixl_manager.shutdown()
-        
-        # Reconstruct full HuggingFace weights from shards
-        logger.info(f"Reconstructing full weights from {len(all_rank_weights)} ranks...")
-        full_weights = self._reconstruct_full_weights(all_rank_weights)
-        
-        # Save as HuggingFace format
-        self._save_hf_format(full_weights)
-        
-        # Store transfer stats
-        self.transfer_stats = {
-            "total_bytes": total_bytes,
-            "total_time": total_time,
-            "bandwidth_gbps": (total_bytes * 8) / (total_time * 1e9) if total_time > 0 else 0,
-        }
-        
-        logger.info(
-            f"Transfer complete: {total_bytes / 1e9:.2f} GB total "
-            f"in {total_time:.2f}s ({self.transfer_stats['bandwidth_gbps']:.1f} Gbps)"
-        )
-    
-    def _reconstruct_full_weights(
-        self, all_rank_weights: dict[int, dict[str, torch.Tensor]]
-    ) -> dict[str, torch.Tensor]:
-        """
-        Reconstruct full HuggingFace weights from TP-sharded weights.
-        
-        Reverses the sharding done by _shard_tensor_for_rank on the source.
-        """
-        # Debug: Log what we have
-        logger.info(f"Reconstruction: {len(all_rank_weights)} ranks")
-        for rank, weights in all_rank_weights.items():
-            logger.info(f"  Rank {rank}: {len(weights)} tensors")
-        
-        if len(all_rank_weights) == 1:
-            # No TP, just return the single rank's weights
-            return list(all_rank_weights.values())[0]
-        
-        # Get all tensor names from rank 0 (or first available rank)
-        first_rank = min(all_rank_weights.keys())
-        rank0_weights = all_rank_weights[first_rank]
-        full_weights = {}
-        
-        if len(rank0_weights) == 0:
-            logger.error("Rank 0 weights are empty!")
-            return {}
-        
-        # Patterns for identifying how tensors were sharded
-        col_parallel_patterns = [
-            "q_proj", "k_proj", "v_proj", "gate_proj", "up_proj",
-            "qkv_proj", "gate_up_proj", "query_key_value",
-        ]
-        row_parallel_patterns = [
-            "o_proj", "down_proj", "out_proj", "dense",
-        ]
-        
-        for name in rank0_weights.keys():
-            # Collect this tensor from all ranks (already on CPU)
-            shards = []
-            for rank in sorted(all_rank_weights.keys()):
-                if name in all_rank_weights[rank]:
-                    shards.append(all_rank_weights[rank][name])
-            
-            if len(shards) == 1:
-                # Not sharded
-                full_weights[name] = shards[0]
-                continue
-            
-            # Determine how to concatenate based on sharding pattern
-            concat_dim = None
-            for pattern in col_parallel_patterns:
-                if pattern in name.lower():
-                    concat_dim = -1  # Was sharded on last dim, concat on last dim
-                    break
-            
-            if concat_dim is None:
-                for pattern in row_parallel_patterns:
-                    if pattern in name.lower():
-                        concat_dim = 0  # Was sharded on first dim, concat on first dim
-                        break
-            
-            if concat_dim is not None:
-                # Concatenate shards
-                full_weights[name] = torch.cat(shards, dim=concat_dim)
-            else:
-                # Not a sharded tensor (e.g., embed_tokens, layer_norm)
-                # All ranks have the same tensor, just use rank 0's
-                full_weights[name] = shards[0]
-        
-        total_size = sum(t.numel() * t.element_size() for t in full_weights.values())
-        logger.info(f"Reconstructed {len(full_weights)} full tensors ({total_size / 1e9:.2f} GB)")
-        
-        return full_weights
-    
-    def _save_hf_format(self, weights: dict[str, torch.Tensor]) -> None:
-        """Save weights in HuggingFace safetensors format."""
-        try:
-            from safetensors.torch import save_file
-        except ImportError:
-            raise ImportError("safetensors required. Install with: pip install safetensors") from None
-        
-        # Save as single model.safetensors file
-        weights_path = self.checkpoint_dir / "model.safetensors"
-        save_file(weights, str(weights_path))
-        
-        total_size = sum(t.numel() * t.element_size() for t in weights.values())
-        logger.info(f"Saved HuggingFace format weights to {weights_path} ({total_size / 1e9:.2f} GB)")
-        
-        # Free memory
-        del weights
-        torch.cuda.empty_cache()
-    
-    def _allocate_tensors(
-        self, tensor_protos, device_id: int
-    ) -> dict[str, torch.Tensor]:
-        """Allocate GPU tensors matching source layout."""
-        weights = {}
-        
-        with torch.cuda.device(device_id):
-            for i, t in enumerate(tensor_protos):
-                dtype = self._parse_dtype(t.dtype)
-                
-                # Debug: Log first few tensors to verify shape is being received
-                if i < 3:
-                    logger.info(f"  Tensor {t.name}: shape={list(t.shape) if t.shape else 'NONE'}, size={t.size}")
-                
-                # Use shape if available, otherwise fall back to flat tensor
-                if t.shape and len(t.shape) > 0:
-                    shape = tuple(t.shape)
-                else:
-                    # Legacy fallback for protos without shape
-                    elem_size = self._dtype_size(dtype)
-                    numel = t.size // elem_size
-                    shape = (numel,)
-                    logger.warning(f"  Tensor {t.name} has no shape, using flat: {shape}")
-                
-                weights[t.name] = torch.empty(
-                    shape, dtype=dtype, device=f"cuda:{device_id}"
-                )
-        
-        return weights
-    
-    def _parse_dtype(self, dtype_str: str) -> torch.dtype:
-        """Parse dtype string to torch.dtype."""
-        dtype_map = {
-            "torch.float16": torch.float16,
-            "torch.float32": torch.float32,
-            "torch.bfloat16": torch.bfloat16,
-            "torch.int8": torch.int8,
-            "torch.int32": torch.int32,
-            "torch.int64": torch.int64,
-        }
-        
-        # Handle float8 types if available
-        if hasattr(torch, "float8_e4m3fn"):
-            dtype_map["torch.float8_e4m3fn"] = torch.float8_e4m3fn
-        if hasattr(torch, "float8_e5m2"):
-            dtype_map["torch.float8_e5m2"] = torch.float8_e5m2
-        
-        return dtype_map.get(dtype_str, torch.float16)
-    
-    def _dtype_size(self, dtype: torch.dtype) -> int:
-        """Get size in bytes for dtype."""
-        return torch.tensor([], dtype=dtype).element_size()
-    
-    def _save_weights(self, weights: dict[str, torch.Tensor], rank: int) -> None:
-        """Save weights to safetensors file."""
-        try:
-            import safetensors.torch
-        except ImportError:
-            raise ImportError("safetensors not installed. Install with: pip install safetensors") from None
-        
-        # Move to CPU for saving
-        cpu_weights = {k: v.cpu() for k, v in weights.items()}
-        
-        weights_path = self.checkpoint_dir / f"rank{rank}.safetensors"
-        safetensors.torch.save_file(cpu_weights, str(weights_path))
-        logger.info(f"Saved rank {rank} weights to {weights_path}")
-        
-        # Free GPU memory
-        del weights
-        torch.cuda.empty_cache()
-    
-    def _build_engine(self) -> str:
-        """Build TRT-LLM engine from checkpoint."""
-        logger.info("Building TRT-LLM engine from checkpoint...")
-        
-        self.engine_dir.mkdir(parents=True, exist_ok=True)
-        
-        # Build command
-        cmd = [
-            "trtllm-build",
-            "--checkpoint_dir", str(self.checkpoint_dir),
-            "--output_dir", str(self.engine_dir),
-        ]
-        
-        # Add build config options
-        for key, value in self.build_config.items():
-            cmd.extend([f"--{key}", str(value)])
-        
-        logger.info(f"Running: {' '.join(cmd)}")
-        
-        t0 = time.perf_counter()
-        result = subprocess.run(
-            cmd,
-            capture_output=True,
-            text=True,
-        )
-        build_time = time.perf_counter() - t0
-        
-        if result.returncode != 0:
-            logger.error(f"trtllm-build failed:\n{result.stderr}")
-            raise RuntimeError(f"trtllm-build failed with code {result.returncode}")
-        
-        logger.info(f"Engine built in {build_time:.1f}s at {self.engine_dir}")
-        
-        return str(self.engine_dir)
-    
-    def get_transfer_stats(self) -> dict:
-        """Get transfer statistics."""
-        return self.transfer_stats
-
-
-def create_trtllm_from_mx(
-    model_name: str,
-    mx_server: str = "modelexpress-server:8001",
-    output_dir: str = "/tmp/mx_trtllm",
-    build_config: dict | None = None,
-    skip_build: bool = False,
-    use_pytorch_backend: bool = False,
-):
-    """
-    Create TRT-LLM model using ModelExpress P2P transfer.
-    
-    This is a convenience function that:
-    1. Receives weights via RDMA from a running source
-    2. For HuggingFace sources: Returns path for PyTorch backend
-    3. For checkpoint sources: Optionally builds TensorRT engine
-    
-    Args:
-        model_name: Model identifier in ModelExpress
-        mx_server: ModelExpress server address
-        output_dir: Local directory for received weights
-        build_config: Optional trtllm-build configuration
-        skip_build: If True, skip engine build (checkpoint mode)
-        use_pytorch_backend: If True, prepare for TRT-LLM PyTorch backend
-        
-    Returns:
-        Path to model directory for TRT-LLM LLM()
-        
-    Example (HuggingFace mode - DeepSeek-V3):
-        # Receive HuggingFace weights
-        model_path = create_trtllm_from_mx(
-            "deepseek-v3",
-            use_pytorch_backend=True
-        )
-        
-        # Use with TRT-LLM PyTorch backend
-        from tensorrt_llm import LLM
-        llm = LLM(model=model_path, backend="pytorch")
-        
-    Example (Checkpoint mode - traditional):
-        # Receive and build engine
-        engine_dir = create_trtllm_from_mx("llama-70b")
-        
-        # Use with TRT-LLM
-        from tensorrt_llm import LLM
-        llm = LLM(model=engine_dir)
-    """
-    loader = MxTrtllmTargetLoader(
-        model_name=model_name,
-        mx_server=mx_server,
-        output_dir=output_dir,
-        build_config=build_config,
-    )
-    
-    return loader.load(skip_build=skip_build, use_pytorch_backend=use_pytorch_backend)
-
-
-# Example usage for DeepSeek-V3
-if __name__ == "__main__":
-    import argparse
-    
-    parser = argparse.ArgumentParser(description="TRT-LLM P2P Transfer Example")
-    parser.add_argument("--mode", choices=["source", "target"], required=True)
-    parser.add_argument("--model-path", required=True, help="HuggingFace model or checkpoint path")
-    parser.add_argument("--model-name", default="deepseek-v3")
-    parser.add_argument("--mx-server", default="modelexpress-server:8001")
-    parser.add_argument("--tp-size", type=int, default=8)
-    parser.add_argument("--output-dir", default="/tmp/mx_trtllm")
-    
-    args = parser.parse_args()
-    
-    if args.mode == "source":
-        # Source: Load HuggingFace model and publish for P2P
-        publisher = MxTrtllmSourcePublisher(
-            model_name=args.model_name,
-            mx_server=args.mx_server,
-            hf_model_path=args.model_path,
-            tp_size=args.tp_size,
-        )
-        publisher.initialize()
-        
-        print(f"Source ready! Model {args.model_name} published for P2P transfer")
-        print("Press Ctrl+C to shutdown...")
-        
-        try:
-            import signal
-            signal.pause()
-        except KeyboardInterrupt:
-            publisher.shutdown()
-    
-    else:
-        # Target: Receive weights and prepare for inference
-        model_path = create_trtllm_from_mx(
-            model_name=args.model_name,
-            mx_server=args.mx_server,
-            output_dir=args.output_dir,
-            use_pytorch_backend=True,
-        )
-        
-        print(f"Model received at: {model_path}")
-        print("Use with TRT-LLM:")
-        print(f"  from tensorrt_llm import LLM")
-        print(f"  llm = LLM(model='{model_path}', backend='pytorch')")
