+++ modelexpress_server/src/metadata_backend.rs
@@ -127,7 +127,6 @@ pub struct TensorRecord {
     pub size: u64,
     pub device_id: u32,
     pub dtype: String,
-    pub shape: Vec<i64>,
 }
 
 // Conversions from gRPC types
@@ -162,7 +161,6 @@ impl From<modelexpress_common::grpc::p2p::TensorDescriptor> for TensorRecord {
             size: desc.size,
             device_id: desc.device_id,
             dtype: desc.dtype,
-            shape: desc.shape,
         }
     }
 }
@@ -203,7 +201,6 @@ impl From<TensorRecord> for modelexpress_common::grpc::p2p::TensorDescriptor {
             size: record.size,
             device_id: record.device_id,
             dtype: record.dtype,
-            shape: record.shape,
         }
     }
 }
