+++ modelexpress_server/src/metadata_backend/kubernetes.rs
@@ -69,7 +69,6 @@ impl KubernetesBackend {
                 size: t.size.to_string(),
                 device_id: t.device_id,
                 dtype: t.dtype.clone(),
-                shape: t.shape.clone(),
             })
             .collect();
 
@@ -160,7 +159,6 @@ impl KubernetesBackend {
                     size,
                     device_id: t.device_id,
                     dtype: t.dtype,
-                    shape: t.shape,
                 })
             })
             .collect::<MetadataResult<Vec<_>>>()?;
